requirements.txt
tweepy==4.14.0
gspread==6.0.0
google-auth==2.26.2
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.2.0
